refactor(web): 機能固有のエラーメッセージ解決を各 feature へ移設
resolveLoginErrorMessage / resolveSignupErrorMessage は login/signup 固有
（401/404→認証情報エラー、409→既存メール）のため各 feature に移し、entities は
両者が使う共通解決 resolveCommonAuthErrorMessage と共有文言に絞る。
