+++ application/apps/web/src/client/entities/auth/index.ts
@@ -1,8 +1,4 @@
-export {
-  AUTH_ERROR_MESSAGES,
-  resolveLoginErrorMessage,
-  resolveSignupErrorMessage,
-} from './lib/error-message'
+export { AUTH_ERROR_MESSAGES, resolveCommonAuthErrorMessage } from './lib/error-message'
 export { resolveTurnstileSiteKey } from './lib/turnstile'
 export { default as useSession, SESSION_SWR_KEY } from './model/use-session'
 export { type AuthenticateErrors, authenticateFormSchema } from './model/validation'
